Added Freebox model.
